requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==25.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.2.1
colorama==0.4.6
Django==5.1.2
django-js-asset==3.1.2
django-mptt==0.17.0
factory_boy==3.3.1
Faker==30.6.0
h11==0.14.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==3.0.1
mypy_extensions==1.1.0
outcome==1.3.0.post0
packaging==24.1
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.5.0
pycparser==2.22
PySocks==1.7.1
pytest==8.3.3
pytest-base-url==2.1.0
pytest-django==4.9.0
pytest-factoryboy==2.7.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-selenium==4.1.0
pytest-variables==3.1.0
python-dateutil==2.9.0.post0
requests==2.32.3
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
tenacity==9.0.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
